third_party/WebKit/LayoutTests/editing/undo/undo-forward-delete-boundary-expected.txt

mac
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "rd "
PASS selection.focusOffset is 2
PASS selection.toString() is "word"
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected only on mac."
win
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This wo"
PASS selection.focusOffset is 5
PASS selection.toString() is ""
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected only on mac."
unix
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This wo"
PASS selection.focusOffset is 5
PASS selection.toString() is ""
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected only on mac."
android
PASS selection.anchorNode.data is "This wo"
PASS selection.anchorOffset is 5
PASS selection.focusNode.data is "This wo"
PASS selection.focusOffset is 5
PASS selection.toString() is ""
PASS $("test").innerHTML is "This wo<b>rd </b>should be selected only on mac."
PASS successfullyParsed is true

TEST COMPLETE
